Update draft-stenberg-homenet-hncp-00.xml
wording, grammar cleanup through end of data model. Only major changes are to be clear that encryption or signing is possible, which implies the need to be able to identify the security algorithm (something that will be required anyway). Some question on whether or not the TLVs MUST be sent in order, or whether they are always reordered to the well-defined order on receipt. The correct answer might be: "Yes, both", in the name of conservative sending and liberal receipt, though this is twice the overall processing. Not specified in the data model for now.